Move simple tests from smartcontractgw to contractstore
